Make listreactions less spammy
